python/requirements.txt
django==2.1.5
django-allauth==0.38.0
psycopg2==2.7.7
gunicorn==19.9.0